requirements.txt
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.137.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
python-dotenv==1.0.1
